rename fraction plotter
