+++ spatialprofilingtoolbox/graphs/importance_fractions.py
